app/services/collector/requirements.txt
Flask==1.0
Flask-Script==2.0.5
Flask-Testing==0.6.2
flask-sqlalchemy==2.3.2
mysqlclient
requests
flask-migrate==2.1.1